Clean up VendorItems component a bit
